SRGAN/requirements.txt
tensorflow-gpu==2.0.0-beta1
matplotlib
Pillow
tqdm
imageio
opencv-python
requests
